Revert mistakes
