+++ base/uk.ac.stfc.isis.ibex.ui.log/src/uk/ac/stfc/isis/ibex/ui/log/widgets/LogDisplay.java
@@ -119,7 +119,6 @@ public class LogDisplay extends Canvas {
 	private MenuItem mnuSaveAll;
 	private MenuItem mnuSaveSelected;
     private AsyncMessageModerator asyncMessageModerator;
-    private SearchControlViewModel searchControlViewModel;
 
     /**
      * Default constructor.
@@ -147,7 +146,7 @@ public class LogDisplay extends Canvas {
      */
 	public void setModel(final LogDisplayModel model) {
 		this.model = model;
-        this.searchControlViewModel.setSearcher(model);
+        // this.searchControl.setSearcher(model);
         this.asyncMessageModerator = new AsyncMessageModerator(model);
 
 		// Listen for updates to the list of messages to be displayed
@@ -309,10 +308,8 @@ public class LogDisplay extends Canvas {
 		// Layout
 		setLayout(new GridLayout(1, false));
 
-        this.searchControlViewModel = new SearchControlViewModel();
-
 		// Add the search box
-        searchControl = new SearchControl(this, model, this.searchControlViewModel);
+		searchControl = new SearchControl(this, model);
 		searchControl.setLayoutData(new GridData(GridData.FILL_HORIZONTAL));
 
         // Add table title label
